+++ src/rpc/net.cpp
@@ -285,6 +285,9 @@ static RPCHelpMan getpeerinfo() {
                         heights.push_back(height);
                     }
                     obj.pushKV("inflight", heights);
+                    obj.pushKV("addr_processed", statestats.m_addr_processed);
+                    obj.pushKV("addr_rate_limited",
+                               statestats.m_addr_rate_limited);
                 }
                 if (IsDeprecatedRPCEnabled(gArgs, "whitelisted")) {
                     // whitelisted is deprecated in v0.24.7 for removal in v0.25