+++ doc/release-notes.md
@@ -8,3 +8,6 @@ This release includes the following features and fixes:
  - Add a `-fixedseeds` option which can be set to 0 to disable the hardcoded seeds.
    This can be used in conjunction with `dsnssed=0` to create a trusted peer only setup.
    In this case the nodes need to be added manually with the `-addnode` option or the `addnode` RPC.
+ - The node will now limit the rate at which the addresses received via p2p messages are processed.
+   This can be bypassed if needed by granting the `addr` permission to a peer(see the `-whitelist`
+   option for details).